feat(cli): implement search interface
