+++ cli/.env.example
@@ -3,9 +3,5 @@ GO_ENV=dev
 # service
 API_URI=http://api:8080/v1
 
-# hotpepper
-HOTPEPPER_API_URL=https://webservice.recruit.co.jp/hotpepper/
-HOTPEPPER_API_KEY=
-
 # line
 LINE_NOTIFY_API_URL=https://notify-api.line.me/api/notify
