+++ cli/api/factory/search/params.go
@@ -1,0 +1,6 @@
+package search
+
+type ListRestaurantsByCityAndGenreParams struct {
+	City  string
+	Genre string
+}
