+++ cli/api/repository/genre/repository.go
@@ -1,8 +1,6 @@
 package genre
 
-import (
-	"github.com/taga3s/pecopeco-cli/api/client/hotpepper"
-)
+import "github.com/taga3s/pecopeco-cli/api/client/app"
 
 type Repository interface {
 	List() (ListResponse, error)
@@ -16,7 +14,7 @@ func New() Repository {
 
 func (r *repository) List() (ListResponse, error) {
 	listResponse := ListResponse{}
-	if err := hotpepper.HttpClient("GET", "/genre/v1/", "&format=json", &listResponse); err != nil {
+	if err := app.HttpClient("GET", "/search/genres", nil, &listResponse); err != nil {
 		return ListResponse{}, err
 	}
 	return listResponse, nil
