+++ cli/api/repository/notify_to_line/repository.go
@@ -1,15 +1,13 @@
-package restaurant
+package notifytoline
 
 import (
 	"fmt"
 
-	"github.com/taga3s/pecopeco-cli/api/client/hotpepper"
 	"github.com/taga3s/pecopeco-cli/api/client/notify"
 	"github.com/taga3s/pecopeco-cli/api/client/util"
 )
 
 type Repository interface {
-	List(request ListRequest) (ListResponse, error)
 	NotifyToLINE(request NotifyToLINERequest) error
 }
 
@@ -17,14 +15,6 @@ type repository struct{}
 
 func New() Repository {
 	return &repository{}
-}
-
-func (r *repository) List(request ListRequest) (ListResponse, error) {
-	listResponse := ListResponse{}
-	if err := hotpepper.HttpClient("GET", "/gourmet/v1/", fmt.Sprintf("&keyword=%s&genre=%s&count=100&format=json", request.City, request.Genre), &listResponse); err != nil {
-		return ListResponse{}, err
-	}
-	return listResponse, nil
 }
 
 func (r *repository) NotifyToLINE(request NotifyToLINERequest) error {
