+++ cli/api/repository/notify_to_line/response.go
@@ -1,0 +1,5 @@
+package notifytoline
+
+type NotifyToLINEResponse struct {
+	Status int `json:"status"`
+}
