+++ cli/api/repository/search/request.go
@@ -1,0 +1,6 @@
+package search
+
+type ListRestaurantsByCityAndGenreRequest struct {
+	City  string
+	Genre string
+}
